+++ src/main/java/com/example/getinline/constant/EventStatus.java
@@ -1,4 +1,4 @@
-package com.example.getinline.domain.constant;
+package com.example.getinline.constant;
 
 public enum EventStatus {
     PENDING, OPENED, CLOSED, CANCELLED, ABORTED
